Refactoring: user app , Add user/validate.py
