+++ students/wooseok-choi/user/models.py
@@ -1,11 +1,11 @@
 from django.db import models
 
-# Create your models here.
+
 class User(models.Model):
-    user_email = models.CharField(max_length = 45)
-    user_password = models.CharField(max_length = 45)
-    phone_number = models.CharField(max_length = 45)
-    nickname = models.CharField(max_length = 45)
+    email        = models.CharField(max_length = 45)
+    password     = models.CharField(max_length = 20)
+    phone_number = models.CharField(max_length = 15)
+    nickname     = models.CharField(max_length = 40)
 
     class Meta:
         db_table = 'users'
