+++ students/wooseok-choi/user/urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
-from user.views import UserView
+from user.views import SignupView
 
 urlpatterns = [
     
-    path('/users', UserView.as_view()),
+    path('/signup', SignupView.as_view()),
 
     ]
